# docs/generated/openapi-gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QAgent API Gateway",
    "description": "\n## QAgent API Gateway\n\nControl-plane API for QAgent — a native Agent Runtime for long-running Agent Teams\n(plan → execute → recover → deliver) with sandboxed tools and observability.\n\n### Features\n\n- **Models Management**: Query and retrieve available AI models\n- **MCP Configuration**: Manage Model Context Protocol (MCP) server configurations\n- **Memory Management**: Access and manage global memory data for personalized conversations\n- **Skills Management**: Query and manage skills and their enabled status\n- **Artifacts**: Access thread artifacts and generated files\n- **Health Monitoring**: System health check endpoints\n\n### Architecture\n\nLangGraph is mounted in-process under `/api/langgraph` (single uvicorn process).\nThis gateway provides custom endpoints for models, MCP configuration, skills, artifacts, and more.\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/channels/{name}/restart": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Restart Channel",
        "description": "Restart a specific IM channel.",
        "operationId": "restart_channel_api_channels__name__restart_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelRestartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/{name}/enable": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Enable Channel",
        "description": "Enable or disable a specific IM channel.",
        "operationId": "enable_channel_api_channels__name__enable_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelEnableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelEnableResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/{name}/config": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Get Channel Config",
        "description": "Get the configuration of a specific IM channel.",
        "operationId": "get_channel_config_api_channels__name__config_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "channels"
        ],
        "summary": "Update Channel Config",
        "description": "Update the configuration of a specific IM channel.",
        "operationId": "update_channel_config_api_channels__name__config_put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Config"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelRestartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/feishu/push": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Feishu Push",
        "description": "Send a proactive Feishu message (markdown card). Intended for cron or other schedulers.\n\nConfigure ``channels.feishu.push_secret`` in ``config.yaml`` or set ``EVOFLOW_FEISHU_PUSH_SECRET``.\nSend the same value in header ``X-QAgent-Feishu-Push-Secret`` or ``Authorization: Bearer <secret>``.\nRequires the Feishu channel to be enabled and running (Gateway started with channel service).",
        "operationId": "feishu_push_api_channels_feishu_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeishuPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeishuPushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/channels/feishu/registration/{session_id}/poll": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Feishu Registration Poll",
        "description": "Poll Feishu registration status.\n\nCall this every 2-3 seconds after ``/begin``.  When ``status == \"completed\"``,\nthe response contains ``app_id`` and ``app_secret``.",
        "operationId": "feishu_registration_poll_api_channels_feishu_registration__session_id__poll_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeishuRegistrationPollResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/feishu/registration/{session_id}/apply": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Feishu Registration Apply",
        "description": "Apply Feishu registration credentials.\n\nSaves ``app_id`` / ``app_secret`` to ``config.yaml`` and automatically starts\n(or restarts) the Feishu channel.  Intended to be called after ``/poll``\nreturns ``status == \"completed\"``.",
        "operationId": "feishu_registration_apply_api_channels_feishu_registration__session_id__apply_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeishuRegistrationApplyRequest",
                "default": {
                  "enabled": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeishuRegistrationApplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelEnableRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "ChannelEnableRequest"
      },
...
      "ChannelRestartResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ChannelRestartResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}